requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt<=5.5.1
psycopg2-binary>=2.9
pgvector>=0.2.4
redis>=5.0
transformers>=4.30
sentence-transformers>=2.2
faiss-cpu>=1.7
langchain>=0.1
peft>=0.4.0
lora>=0.1
qlora>=0.0.1
python-dotenv>=1.0
requests>=2.31
uvicorn>=0.23
django-redis
drf-extensions
djangorestframework-ratelimit
